refactor: Article API 도메인으로 이동으로 인한 기존 파일 경로 수정
