+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/domain/entity/CommentEntity.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.entity;
+package com.icehufs.icebreaker.domain.article.domain.entity;
 
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
@@ -11,8 +11,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
-import com.icehufs.icebreaker.dto.request.article.PatchCommentRequestDto;
-import com.icehufs.icebreaker.dto.request.article.PostCommentRequestDto;
+import com.icehufs.icebreaker.domain.article.dto.request.PatchCommentRequestDto;
+import com.icehufs.icebreaker.domain.article.dto.request.PostCommentRequestDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
