+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/dto/response/DeleteCommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.response.article;
+package com.icehufs.icebreaker.domain.article.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
