+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/repository/ArtileListViewRepository.java
@@ -1,11 +1,11 @@
-package com.icehufs.icebreaker.repository;
+package com.icehufs.icebreaker.domain.article.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.icehufs.icebreaker.entity.Article;
+import com.icehufs.icebreaker.domain.article.domain.entity.Article;
 
 @Repository
 public interface ArtileListViewRepository extends JpaRepository<Article, Integer>{
