+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/repository/CommentRepository.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.repository;
+package com.icehufs.icebreaker.domain.article.repository;
 
 import java.util.List;
 
@@ -7,7 +7,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.icehufs.icebreaker.entity.CommentEntity;
+import com.icehufs.icebreaker.domain.article.domain.entity.CommentEntity;
 
 
 @Repository
