+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/repository/FavoriteRepository.java
@@ -1,11 +1,11 @@
-package com.icehufs.icebreaker.repository;
+package com.icehufs.icebreaker.domain.article.repository;
 
 import jakarta.transaction.Transactional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.icehufs.icebreaker.entity.FavoriteEntity;
+import com.icehufs.icebreaker.domain.article.domain.entity.FavoriteEntity;
 import com.icehufs.icebreaker.entity.primaryKey.FavoritePk;
 
 @Repository
