+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/service/implement/ArticleServiceImplement.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.service.implement;
+package com.icehufs.icebreaker.domain.article.service.implement;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,19 +6,36 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import com.icehufs.icebreaker.dto.request.article.*;
+import com.icehufs.icebreaker.domain.article.dto.request.PatchArticleRequestDto;
+import com.icehufs.icebreaker.domain.article.dto.request.PatchCommentRequestDto;
+import com.icehufs.icebreaker.domain.article.dto.request.PostArticleRequestDto;
+import com.icehufs.icebreaker.domain.article.dto.request.PostCommentRequestDto;
+import com.icehufs.icebreaker.domain.article.dto.response.CheckArticleFavoriteResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.CheckOwnOfArticleResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.DeleteArticleAdminResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.DeleteArticleResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.DeleteCommentResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.GetArticleListResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.GetArticleResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.GetCommentListResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.GetUserArticleListResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.PatchArticleResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.PatchCommentResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.PostArticleResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.PostCommentResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.PutFavoriteResponseDto;
+import com.icehufs.icebreaker.domain.article.dto.response.PutResolvedArticleResponseDto;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
-import com.icehufs.icebreaker.dto.response.article.*;
-import com.icehufs.icebreaker.entity.Article;
-import com.icehufs.icebreaker.entity.ArticleCategoryEnum;
-import com.icehufs.icebreaker.entity.CommentEntity;
-import com.icehufs.icebreaker.entity.FavoriteEntity;
-import com.icehufs.icebreaker.repository.ArticleRepository;
-import com.icehufs.icebreaker.repository.ArtileListViewRepository;
-import com.icehufs.icebreaker.repository.CommentRepository;
-import com.icehufs.icebreaker.repository.FavoriteRepository;
+import com.icehufs.icebreaker.domain.article.domain.entity.Article;
+import com.icehufs.icebreaker.domain.article.domain.entity.ArticleCategoryEnum;
+import com.icehufs.icebreaker.domain.article.domain.entity.CommentEntity;
+import com.icehufs.icebreaker.domain.article.domain.entity.FavoriteEntity;
+import com.icehufs.icebreaker.domain.article.repository.ArticleRepository;
+import com.icehufs.icebreaker.domain.article.repository.ArtileListViewRepository;
+import com.icehufs.icebreaker.domain.article.repository.CommentRepository;
+import com.icehufs.icebreaker.domain.article.repository.FavoriteRepository;
 import com.icehufs.icebreaker.repository.UserRepository;
-import com.icehufs.icebreaker.service.ArticleService;
+import com.icehufs.icebreaker.domain.article.service.ArticleService;
 import com.icehufs.icebreaker.util.EncryptionUtil;
 
 import lombok.RequiredArgsConstructor;
