+++ backend/src/main/java/com/icehufs/icebreaker/dto/object/ArticleListItem.java
@@ -4,8 +4,8 @@ package com.icehufs.icebreaker.dto.object;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.icehufs.icebreaker.entity.Article;
-import com.icehufs.icebreaker.entity.ArticleCategoryEnum;
+import com.icehufs.icebreaker.domain.article.domain.entity.Article;
+import com.icehufs.icebreaker.domain.article.domain.entity.ArticleCategoryEnum;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
